third_party/blink/web_tests/flag-specific/highdpi/fast/forms/color/color-picker-appearance-color-well-keyboard-navigation-after-drag-expected.txt

This is a testharness.js-based test.
FAIL Color picker: Moving the color well's selection ring using the keyboard after a drag. assert_equals: expected "123" but got "119"
Harness: the test ran to completion.
